make JavaObject, ArrayElement 'static
This simplifies some GAT lifetime bounds the compiler is too silly
to figure out. It makes sense because JavaObject is meant to represent
Java types and they don't have lifetimes. However, if desired we could
revert it, it would just make NotNull a bit more complex to write
(we'd need to use a "Nullable" trait implemented by `Self::Output<'jvm>`
or something like that.)