Pull out `mod java`, add `Throwable`
Create a new top-level `java.rs` that contains the JDK core class declarations
that are shared across `duchess` invocations.

We may eventually want to move `struct JavaArray` to sit outside of `mod java` to
not confuse it with the real `java.lang.Arrays` or `java.lang.reflect.Array` classes
that the JDK provides. Java arrays directly inherit from `Object`, which is confusing!

This commit leaves the temporary `collections` classes in place to not break the
examples, but they'll soon be autogenerated.

With these declarations in one place, this adds temporary `Throwable` and
`StackTraceElement` `javap` dumps to unblock throw and catch support.