Cache MethodID for constructor & method calls
Use the same `OnceCell` caching mechanism for looking up method IDs
for constructor and method calls. This forces us to switch to the
`*_unchecked()` JNI functions. The checks in the regular versions
should be redundant with our own type system. It also avoids
parsing the method signature string on each call!