Fix thrown exceptions being buried during execute() calls
A Duchess user can execute most `crate::JvmOp`s at any time inside of a
`crate::Jvm::with()` call. The returned `Result` should correctly
contain an `Error::Thrown` if a Java exception was thrown. If we waited
to materialize errors until exiting `with` or in a `catching()` block, the
execute call would return an internal `Error::Jni` instead and the
exception would be buried.

`Error::Jni` should be reserved for unexpected JNI failures and not
exposed to Duchess users for normal scenarios like a thrown exception.