make Rust conversion a JVM op
So now you do `foo.to_rust().execute(jvm)` and not `foo.into_rust(jvm)`.
This will allow us to compose it with other things, e.g., we will be
able to do rust conversion before we start catching errors.

It is also based on a trait (`ToRust`) that we will be able to derive.