change to `ToRust<R>` instead of assoc type
This is more flexible because it allows a single Java type to
be converted to multiple Rust types. It may mean more annotation
on the user's part, but they can write `.to_rust::<RustType>()`
to disambiguate if needed (the compiler uses the impls otherwise).

The goal here is to allow us to auto-derive multiple `ToRust<E>`
impl for some Rust error type E for `Throwable`.