Generate test class files via build.rs
Add a custom build script to detect changes and auto-build class files for
Java tests and examples. We assume that javap is already in the path, so we
think it's safe to also assume javac is available. Also update the cargo
CLASSPATH to point to the new generated directory and clean up the
previously checked-in class files.

Fixes #88