Add in checks for the type of Java Class/Interface
Works for both fully reflected classes and for specified classes.
Added UI tests for both cases, and updated other portions where needed
to specify the correct type of Java import.

For Reflected classes, fixed an error in specifying the span and kind
of the class. Previously, they took the span of the reflection cache,
which meant that the error could confusingly point to the main
`java_package!` macro invocation.