Add in checks for the type of Java Class/Interface (#168)
* Add in checks for the type of Java Class/Interface

Works for both fully reflected classes and for specified classes.
Added UI tests for both cases, and updated other portions where needed
to specify the correct type of Java import.

For Reflected classes, fixed an error in specifying the span and kind
of the class. Previously, they took the span of the reflection cache,
which meant that the error could confusingly point to the main
`java_package!` macro invocation.

* Fix the casing of the correct class kind

Now in lowercase instead of title case, to more accurately line up with
what users of the library need to provide.

* Bump version to 0.3.0

Extra checks distinguishing interface / class are not backwards
compatible, so a version bump is necessary.