Add shortcuts `exactly_X` for pinning the JNI version
Allows for pinning the JNI version. Should only be used in exact cases,
as this will block off other crates in the tree wanting to use other
versions, if they are compatible. Used in the tests, because the tests
are there to test that the JNI versions still function correctly.