Simplify JNI Version specification based on discussion
No need for complicated `at_least` or `at_most`, as that usually isn't
something that comes up.

Default JNI to 1.6 on Android, and 1.8 for all other targets, but only
have a feature for each JNI version. Duchess will take the highest
version specified in the features, but fail out if not 1.6 in Android
(as Android does not support JNI > 1.6 at time of commit). This fail is
a compile error.

JNI version is now a matrix option in the Rust build workflow.
This runs on every commit and PR.