Allow to specify minimum or maximum JNI versions (#180)
* Allow to specify minimum or maximum JNI versions

Currently, can pick between JNI 1.6 or 1.8. Resolves to one JNI version.

Will fail to compile if users ask for conflicting JNI versions, but will
try and resolve to a JNI version that everyone in the compile tree is
happy with. Also updated the github workflows to run against JNI 1.8
and JNI 1.6. They are done in sets of `at_most` and `at_least`, as rust
features are meant to be additive.

Will fail to compile if `jni_at_most_1_6` and `jni_at_least_1_8` is
requested, which is normal and expected.

* Add shortcuts `exactly_X` for pinning the JNI version

Allows for pinning the JNI version. Should only be used in exact cases,
as this will block off other crates in the tree wanting to use other
versions, if they are compatible. Used in the tests, because the tests
are there to test that the JNI versions still function correctly.

* Simplify JNI Version specification based on discussion

No need for complicated `at_least` or `at_most`, as that usually isn't
something that comes up.

Default JNI to 1.6 on Android, and 1.8 for all other targets, but only
have a feature for each JNI version. Duchess will take the highest
version specified in the features, but fail out if not 1.6 in Android
(as Android does not support JNI > 1.6 at time of commit). This fail is
a compile error.

JNI version is now a matrix option in the Rust build workflow.
This runs on every commit and PR.