pass in "token tree" description of java type
We used to process types in many ways and pass
in a number of various. We now generate most of
those in declarative macro-rules logic in
`java_types.rs` so that we can just pass in
token tree description of the java type and
let the rest of the code work.