Apply jdk fix to 0.3.0 (#202)
* Fix issue on aarch64 with jdk-17.0.15+6-LTS

There seems to be some issue with launch_or_use_existing where it will
panic in the try_launch branch reliably on aarch64 linux hosts
(reproduced on AL2 & AL2023). Adding this block in fixes it.

Note this is the minimal example to reproduce the crash:
```rust
use jni::JavaVM;

fn main() {
    let args = jni::InitArgsBuilder::new()
        .build()
        .expect("Failed to build JVM InitArgs");
    let jvm = JavaVM::new(args).expect("Failed to build JVM");
    duchess::Jvm::builder().launch_or_use_existing().unwrap();
}
```

This patch reliably fixes the panic, not sure why right now!

* Add repro test

Adds a repro test, currently can't get it to fail when commenting out
the fix.

* Update src/jvm.rs

Co-authored-by: Michael Bower <michaelb072@gmail.com>

* Rename use_existing_works.rs to duchess_can_use_jni_launched_jvm.rs

---------

Co-authored-by: Michael Bower <michaelb072@gmail.com>